+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## Unpublished
+### Added
+- button and modal for sending reminder to members of a sponsoring contract
 
 ## v0.7.18 - 2024-09-12
 ### Added
@@ -10,7 +12,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - no double separator lines on dashboard cards
 - show 404 instead of exception when file is missing
-- forgotton membergroup change for event calendar
+- forgotten membergroup change for event calendar
 - null saved in case of no token in event ics controller
 
 ## v0.7.17 - 2024-09-11